Basic platforming in place.
The basic platforming controls are there, and can partially be tweaked. Player object must have PlayerScript and Controller2D scripts as added componects. Any objects it needs to collide with will need to be in the Layer "Environment", this setting is at the Collision mask found on when of the scripts in the Inspector when the Player is selected.